Move scopes into access_control, adjust imports
